+++ CHANGELOG.md
@@ -7,6 +7,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 - Update default color space clipping and expose QPs for tuning [#5561](https://github.com/raster-foundry/raster-foundry/pull/5561)
 
+### Fixed
+- Fix campaign label class summary endpoint [#5560](https://github.com/raster-foundry/raster-foundry/pull/5560)
+- Changed projection initialization so as not to flip xy and move Miami to Guyana [#5563](https://github.com/raster-foundry/raster-foundry/pull/5563)
+
+### Added
+- Made campaign and annotation project delete async [#5559](https://github.com/raster-foundry/raster-foundry/pull/5559)
+
 ## [1.61.2] - 2021-03-30
 ### Changed
 - Campaign creates can include an owner [#5558](https://github.com/raster-foundry/raster-foundry/pull/5558)
@@ -21,6 +28,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Changed
 - Updated setup instructions and scripts [#5544](https://github.com/raster-foundry/raster-foundry/pull/5544)
+- Skip cogification on uploads if the tiff is already a cog [#557](https://github.com/raster-foundry/raster-foundry/pull/5557)
 
 ## [1.61.0] - 2021-03-04
 ### Changed